requirements.txt
alembic==1.7.7
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
configparser==3.5.0
coverage==4.4.2
cryptography==36.0.2
dnspython==2.2.1
email-validator==1.1.3
enum34==1.1.6
flake8==3.3.0
Flask==1.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Mandrill==0.3
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==1.0.1
funcsigs==1.0.2
gunicorn==19.7.1
idna==2.9
importlib-metadata==1.3.0
importlib-resources==5.4.0
iniconfig==1.1.1
install==1.3.5
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.6
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
more-itertools==8.6.0
packaging==20.8
pbr==3.1.1
pep8==1.7.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.3.1
pycparser==2.21
pyflakes==1.5.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pytest==6.2.2
python-decouple==3.6
python-dotenv==0.20.0
python-editor==1.0.3
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
wincertstore==0.2
WTForms==2.3.3
zipp==3.4.0
